+++ apps/domain-clients-example/src/components/cosmos/index.ts
@@ -1,2 +1,3 @@
 export * from './NeutronRead';
 export * from './NeutronTestnetWrite';
+export * from './ConnectCosmosChain';