+++ packages/domain-modal-react/src/index.ts
@@ -1,3 +1,3 @@
-export * from './ui/DomainModalProvider';
 export * from './ui/util';
 export * from './hooks/';
+export * from './ui/context';